coverage: Extract module `mapgen::unused` for handling unused functions
